+++ builder.js
@@ -30,6 +30,7 @@ module.exports = (env) => {
         './src/ts/compounds-table.jsx',
         './src/ts/users-table.jsx',
         './src/ts/jspreadsheet.jsx',
+        './src/ts/jspreadsheet.utils.ts',
         './src/ts/spreadsheet-editor.jsx',
         './src/ts/tags.ts',
         './src/ts/admin.ts',
