+++ src/ts/uploads.ts
@@ -9,6 +9,7 @@ import $ from 'jquery';
 import { Action as MalleAction, Malle } from '@deltablot/malle';
 import '@fancyapps/fancybox/dist/jquery.fancybox.js';
 import { Action, Model } from './interfaces';
+import { loadInJSSpreadsheet } from "./jspreadsheet.utils";
 import { ensureTogglableSectionIsOpen, relativeMoment, reloadElements } from './misc';
 import { displayPlasmidViewer } from './ove';
 import { displayMoleculeViewer, get3dmol } from './3dmol';
@@ -146,9 +147,9 @@ const clickHandler = async (event: Event) => {
 
   // LOAD SPREADSHEET FILE
   } else if (el.matches('[data-action="xls-load-file"]')) {
-    await SpreadsheetEditorHelperC.loadInSpreadsheetEditor(el.dataset.link, el.dataset.name, Number(el.dataset.uploadid));
-    ensureTogglableSectionIsOpen('sheetEditorIcon', 'spreadsheetEditorDiv');
-    document.getElementById('spreadsheetEditor')?.scrollIntoView({ behavior: 'smooth'});
+    await loadInJSSpreadsheet(el.dataset.link, el.dataset.name, Number(el.dataset.uploadid));
+    ensureTogglableSectionIsOpen('sheetEditorIcon', 'jspreadsheetDiv');
+    document.getElementById('jspreadsheet')?.scrollIntoView({ behavior: 'smooth'});
 
   // ARCHIVE UPLOAD
   } else if (el.matches('[data-action="archive-upload"]')) {
